Decouple tests for binary elementwise functions to a dedicated scope
